+++ src/imio/smartweb/core/browser/static/src/edit.js
@@ -5,7 +5,9 @@ jQuery(document).ready(function ($) {
     $(".hide-in-preview, #section-byline, #global_statusmessage").toggle("fast");
     e.preventDefault();
   });
+});
 
+jQuery(window).on("load", function(e) {
   // Move authentic sources menu just before user/personaltools-menulink in Plone toolbar
   var auth_sources = $("#plone-authentic-sources-menu").wrap("<ul class='plonetoolbar-authentic-sources-menu'>").parent();
   $(".personaltools-wrapper").prepend(auth_sources);
@@ -14,4 +16,3 @@ jQuery(document).ready(function ($) {
   var smartweb_help = $("#plone-smartweb-help-menu").wrap("<ul class='plonetoolbar-smartweb-help-menu'>").parent();
   $(".personaltools-wrapper").prepend(smartweb_help);
 });
-